+++ app/models/hacker.rb
@@ -1,110 +1,79 @@
-# User User model, that saves the Github hacker data to a Mongoid doc.
-# @author Islam Wazery <wazery@ubuntu.com>
-# @author Mohamed Yossry <mohamedyosry3000@gmail.com>
+# Hacker model, that saves the Github hacker data to a Mongoid document.
 class Hacker
   include Mongoid::Document
   include Mongoid::Timestamps
-  devise :registerable, :rememberable, :trackable,
-         :omniauthable
+  include Github::Auth
 
   field :email, type: String, default: ''
 
-  ## Rememberable
+  # Rememberable
   field :remember_created_at, type: Time
 
-  ## Trackable
+  # Trackable
   field :sign_in_count,      type: Integer, default: 0
   field :current_sign_in_at, type: Time
   field :last_sign_in_at,    type: Time
   field :current_sign_in_ip, type: String
   field :last_sign_in_ip,    type: String
 
-  ## Omniauthable
-  # field :provider,           type: String, default: ''
-  # field :uid,                type: String, default: ''
+  # Omniauthable
   field :name,               type: String, default: ''
   field :github_token,       type: String, default: ''
+  field :github_uid, type: String
 
   index({ email: 1 }, unique: true, background: true)
-  # index({ uid: 1 },          unique: true, background: true)
   index({ github_token: 1 }, unique: true, background: true)
+  index({ api_secret: 1 }, unique: true, background: true)
+
+  # Github Data
+  field :raw_data, type: Hash
+  field :company, type: String
+  field :username, type: String
+  field :api_secret, type: String
+  field :avatar_url, type: String
+  field :display_name, type: String
+  # TODO: change to total watched repos
+  field :total_watched_repos, type: Integer
+  field :public_gists, type: Integer
+  field :current_scope, type: String # '' or 'user, repo'
+  field :private_gists, type: Integer
+  # TODO: change to total watcehd repos
+  field :total_private_repos, type: Integer
+  field :owned_private_repos, type: Integer
+  field :private_app_github_access_token, type: String
+  field :public_app_github_access_token, type: String
 
   # Validations
   VALID_EMAIL_REGEX = /\A[\w+\-.]+@[a-z\d\-.]+\.[a-z]+\z/i
   validates :email, presence: true, uniqueness: true, format: { with: VALID_EMAIL_REGEX, message: 'is not valid email' }
   validates :name, presence: true
+  validates :avatar_url, presence: true
+  validates :display_name, presence: true
   validates :github_token, presence: true, uniqueness: true
+  validates :api_secret, presence: true, uniqueness: true
 
-  # Class Attributes
-  cattr_writer :octokit
-  cattr_writer :octokit_client
+  # Callbacks
+  after_initialize :_set_api_secret
 
-  def self.octokit
-    @@octokit || Octokit
-  end
-
-  def self.octokit_client
-    @@octokit_client || Octokit::Client
-  end
-
-  def self.github_client(access_token)
-    octokit_client.new(access_token: access_token)
-  end
-
-  def self.sign_up(data)
-    Hacker.create(data)
-    # TODO: Implement Intercom Notifier
-    # IntercomNotifier.perform_in(1.hours, :signed_up_an_hour_ago, hacker.id)
-  end
-
-  def self.find_or_create_by_github_oauth_code(code)
-    access_token = fetch_github_access_token(code)
-
-    hacker = find_by_github_access_token(access_token)
-    return hacker if hacker
-
-    hacker = sign_up(info_by_github_access_token(access_token))
-    hacker
-  end
-
-  def self.find_by_github_access_token(access_token)
-    where(github_token: access_token).first
-  end
-
-  def self.info_by_github_access_token(access_token)
-    client = github_client(access_token)
-    hacker   = client.hacker
-
-    # cached_emails = REDIS.get(github_emails_cache_key(access_token))
-    # emails = JSON.parse(cached_emails) if cached_emails
-    email = client.emails.first[:email] # unless cached_emails
-    # TODO: Check why reject those emails?
-    # emails = emails.reject { |email| email =~ /@hackers\.noreply\.github\.com$/ }
-
-    name = hacker.name.blank? ? email.partition('@').first : hacker.name
-
-    { email: email, name: name, github_token: access_token }
-  end
-
-  def sign_in_formatted
-    {
-      id: id,
-      token: github_token,
-      email: email,
-      name: name,
-      sign_in_count: sign_in_count,
-      created_at: created_at
-    }
+  class << self
+    # Creates the actual hacker into DB.
+    #
+    # @param data [Hash] the data needed to signup the hacker
+    # @return [Hacker] a signed up hacker.
+    def sign_up(data)
+      Hacker.create(data)
+      # TODO: Implement Intercom Notifier
+      # IntercomNotifier.perform_in(1.hours, :signed_up_an_hour_ago, hacker.id)
+    end
   end
 
   private
 
-  def fetch_github_access_token(code)
-    ret = octokit.exchange_code_for_token(code)
+  def _set_api_secret
+    self.api_secret ||= JWTToken.generate
+  end
 
-    return ret[:access_token] if ret[:access_token]
-    # TODO: Handling exception
-    fail ret[:error_description] if ret[:error]
-    fail 'Falied fetching access token'
+  def private_access_granted?
+    current_hacker.private_app_github_access_token.present?
   end
 end